Support writing port mapping to a file
When kubectl port-forward has fully established its port forwarding
connections, write a json mapping of local to remote ports to stdout or
a file if the --ports-file argument is specified.

If the ports file is not "-" for stdout it will be closed once the port
mapping is written. By passing a pipe, the calling program can detect
EOF on the pipe to determine when port-forwarding is fully established.

This is a WIP:

* I'm not super happy with how port-file writing is synchronized,
  and might want to move the actual port-forward invocation into
  a goroutine then synchronize on it, so the port-file writing can
  be called from the main RunPortForward() function body once the
  forwarding is fully up.

* The test structure I've introduced is a bit tortured and really needs
  to be flattened and simplified.

* It bundles an unrelated change to fix stdout/stderr handling
  in the command so that Cobra command error output properly
  respects the genericclioptions streams instead of printing
  usage messages etc to raw stderr.

It's an attempt to fix https://github.com/kubernetes/kubectl/issues/1190